Disable double buffering when DWM is enabled
Windows Vista and newer can use compositing (DWM) which is enabled by default.
It prevents tearing, so if the DWM is enabled, use single buffer. Otherwise on
AMD and nVidia GPU on fullscreen the toolbar and context menu are invisible.

VSync doesn't work when single buffer is chosen, so DWM must prevents tearing.
This is not 3D game, so it should be enough.

Since Windows 8, the DWM is probably always enabled.