Remove old OpenGL module
New OpenGL 2 module is better, because it can be compiled with OpenGL|ES.
The new module doesn't work with old GPU/drivers. Multitexturing, shaders
and non-power-of-two textures support is mandatory! If these features are
not exists, OpenGL won't work and another plugin will be used.

For old GPU XVideo/DirectDraw should be used. QPainter also can be used,
but it doesn't use GPU, so video will tear without compositing.