use QByteArray::constData() instead of QByteArray::data()
Common practice for read-only access; the Qt documentation claims
this is (can be) faster because no data is ever (deep) copied.
The patch modifies the locations where the potential gain should be
the most beneficial.
There is 1 other location PortAudioWriter::writeStream(); that hunk
intersects with another patch I plan to propose and will thus be part
of the corresponding PR (which doesn't mean it cannot be submitted along
with these changes).

Committed from host : Portia.local