Quick fix for the current YouTube playback/download problem.
This modifies the `youtube-dl -g -j` parsing so a list of actual
content streams is obtained, sorted to favour the higher qualities
(read: larger vertical resolution).

It is a quick fix because it will not provide the user with a collapsable
list of resolutions under the video thumbnail. Instead the video will
be treated as content with multi streams (and downloaded to a mkv
container); the user can select video and/or audio streams at will via
the Information widget. Not the most convenient (user preferences for
resolution are ignored) but infinitely better than getting an error on
trying to open an URL that is not a media stream.
The resulting address (as would be reported in the Information widget)
is not one you can do anything with, so the YouTube page address is
reported instead.

TODO: exclude stream URLs that cannot be played anyway and which will
presumably increase processing overhead before the streams starts
playing.

Fixes: #231
Committed from host : Bola