Support reading lyrics meta-data from ID3v2 tags
ID3v2 tags support specifying a description and the language of lyrics.
There can be one lyric per description/language combination. Hence
libavformat makes up keys like `lyrics-foo-eng` in the metadata dictionary
for those fields (see code in `libavformat/id3v2.c` in the FFmpeg
repository or call `ffprobe` on such a file). Even if the description and
language are empty the field name for lyrics becomes `lyrics-`. This change
allows QMPlay2 to display the first of these fields.